Reduce G0 string probe traffic with fingerprints

+++ server_entry/project/string/string_registry.hpp
@@ -202,7 +202,10 @@ private:
 
     std::vector<string_registry::string_record> added_records;
     std::vector<char> added_bytes;
-    std::vector<std::uint32_t> added_index;
+    // Update-local open-addressing buckets pack a 32-bit hash fingerprint with
+    // the one-based local added-record coordinate. Most failed probes therefore
+    // avoid touching added_records and added_bytes.
+    std::vector<std::uint64_t> added_index;
 
     std::vector<string_registry::byte_block> rebuilt_blocks;
     std::vector<string_registry::string_record> rebuilt_records;
